Move backend utils from core app to netbox
