+++ netbox/core/filtersets.py
@@ -4,9 +4,9 @@ from django.utils.translation import gettext as _
 import django_filters
 
 from netbox.filtersets import BaseFilterSet, ChangeLoggedModelFilterSet, NetBoxModelFilterSet
+from netbox.utils import get_data_backend_choices
 from .choices import *
 from .models import *
-from .utils import get_data_backend_choices
 
 __all__ = (
     'DataFileFilterSet',
