+++ netbox/netbox/utils.py
